Added ConfirmDialog component into ui-shared-lite
