+++ modules/channel-web/assets/default.css
@@ -4,6 +4,11 @@
   --main-bg-color-hover: rgba(255, 255, 255, 0.9);
   --main-fg-color: #000000;
   --main-fg-color-hover: rgba(0, 0, 0, 0.07);
+
+  --seashell: #e2e2e2;
+  --lighthouse: #f66f48;
+  --focus-lighthouse: #e06542;
+  --hover-lighthouse: #f39c82;
 }
 
 body {
