+++ src/bp/ui-shared-lite/Dialog/style.scss.d.ts
@@ -1,0 +1,9 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'dialog': string;
+  'dialogBody': string;
+  'form': string;
+}
+declare var cssExports: CssExports;
+export = cssExports;
