+++ src/bp/ui-shared-lite/Dialog/typings.d.ts
@@ -1,0 +1,10 @@
+// @ts-nocheck
+import { IDialogProps } from '@blueprintjs/core'
+
+export interface DialogProps extends IDialogProps {
+  onSubmit?: () => void
+  children: any
+  isOpen?: boolean
+  height?: number
+  size?: 'sm' | 'md' | 'lg'
+}
